Fixed bid validation problem.  Upgraded IntelliJ and Grails.

+++ grails-app/controllers/bitterbidder/ListingController.groovy
@@ -148,7 +148,7 @@ class ListingController {
 
         //this id is for the listing we're bidding upon
         def id = params.int('id')
-        def amt = params.int('amount')
+        def amt = params.float('amount')
         def jsonMap = [id];
 
         if(id){
